+++ changelog.md
@@ -14,7 +14,10 @@
         * Change methods to static in JwtUtils.java
     * Changed login page layout [Pull Request #90](https://github.com/getheimdall/heimdall/pull/90)
     * Adding scroll in tables and update list of the environments [Pull Request #92](https://github.com/getheimdall/heimdall/pull/92)
-
+    * Allow creation of one app with informed clientId [Pull Request #89](https://github.com/getheimdall/heimdall/pull/89)
+        * Create new class AppPersist
+        * Create tests
+    
 * **Bugfix**
 
     * Remove header method from middleware helper class was not working. [Pull Request #76](https://github.com/getheimdall/heimdall/pull/76)